move glassnode to datasource
